Completely optimizes feature engineering module for the data.
Previously was using panda.apply functions to create new features for the data.
The bottleneck here was that the apply functions was being performed on every single row of the dataframe.
This resulted in redundant duplications and a significant decrease in time efficency for preparing the data.

The revamped model creates multiple dataframe and utilizes join and merge methods from the pandas library.
For instance, we create a separate data frame for the team performances, with every row being a unique entry of data that occurred (aka split the original row into two for each time).
We then utilize group by functions on the new data frame (group by team_id and match_date) and aggregate functions that calculate averages up to that point in time.
After the new data frames are created, we can simply merge them back into the original data frame and we are done for all entries for that feature engineering.
This SIGNIFICANTLY reduces the time complexity of the data operations.