Enhances scraping endpoint to update the models after new data is aquired
Logical modification, every time new match data is made available,
the models will be updated to reflect the new data.

train.py module in the ML package is modified to reflect this change.
Now the models can be trained via the endpoint route and also
as a standalone script in the machine_learning root file.